% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova.R
\name{blindDataAnova}
\alias{blindDataAnova}
\title{Suitable for ANOVA:
The main function blindDataAnova takes an input dataset and offers data 
blinding techniques to blind the data: (1) AddNoise, (2) AddBias, 
(3) CreateNew, (4) ScrambleOutcome, and (5) ScramblePredictors.}
\usage{
blindDataAnova(
  df_original,
  y,
  predictors,
  blinding_method = c("AverageWithNoisePerCase", "AverageWithNoisePerGroup",
    "CenterGroups", "MaskGroups", "ScrambleCells", "ScrambleGroups", "ScrambleOutcome",
    "ScramblePredictors"),
  n_permutations = 6,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{blinding_method}{# name of blinding method}

\item{n_permutations}{# number of requested permuted datasets}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The main function blindDataAnova takes an input dataset and offers data 
blinding techniques to blind the data: (1) AddNoise, (2) AddBias, 
(3) CreateNew, (4) ScrambleOutcome, and (5) ScramblePredictors.
}
\examples{
blindDataAnova(df_original = df_sim_anova, 
y = "score",
predictors = c("expert", "conflict"),
blinding_method = "AverageWithNoisePerCase"
) 
 
}
\keyword{ANOVA}
\keyword{blindData,}
