% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_AverageWithNoisePerCase.R
\name{blindDataAnova_AverageWithNoisePerCase}
\alias{blindDataAnova_AverageWithNoisePerCase}
\title{Suitable for ANOVA:
The function AverageWithNoisePerCase averages: each original dependent 
variable value (Y) and random noise that is drawn for each CASE from a 
distribution that is based on the mean(Y) and sd(Y).}
\usage{
blindDataAnova_AverageWithNoisePerCase(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function AverageWithNoisePerCase averages: each original dependent 
variable value (Y) and random noise that is drawn for each CASE from a 
distribution that is based on the mean(Y) and sd(Y).
}
\keyword{Average}
\keyword{Regression}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{with}
