% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_CenterGroups.R
\name{blindDataAnova_CenterGroups}
\alias{blindDataAnova_CenterGroups}
\title{Suitable for ANOVA:
The function CenterGroups deducts, for each group, the average dependent 
variable value (Y) #' from each original dependent variable value (Y)}
\usage{
blindDataAnova_CenterGroups(df_original, y, predictors, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function CenterGroups deducts, for each group, the average dependent 
variable value (Y) #' from each original dependent variable value (Y)
}
\keyword{ANOVA}
\keyword{Center}
\keyword{groups,}
