% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_MaskGroups.R
\name{blindDataAnova_MaskGroups}
\alias{blindDataAnova_MaskGroups}
\title{Suitable for ANOVA:
The function "blindDataAnova_MaskGroups" masks the label names for the 
independent variables.}
\usage{
blindDataAnova_MaskGroups(df_original, y, predictors, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function "blindDataAnova_MaskGroups" masks the label names for the 
independent variables.
}
\keyword{ANOVA}
\keyword{Mask}
\keyword{groups,}
