% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_ScrambleCells.R
\name{blindDataAnova_ScrambleCells}
\alias{blindDataAnova_ScrambleCells}
\title{Suitable for ANOVA:
The function "blindDataAnova_ScrambleCellss" re-assigns at random the labels 
related to the specific combination of groups a case belongs to (e.g., 
expert: low, conflict: no, can get assigned to expert: high, conflict: yes). 
The number of permutations can be requested by user (e.g., if you want to 
have 6 datasets, input n_pertubations = 6; also defaults to 6). This function
works with unequal groups. The function returns a list with dataframes and a 
dataframe that contains all the permutations (labelled by permutation) for 
ease of use. The function also returns the number of possible permutations 
and the number of unique group combinations.}
\usage{
blindDataAnova_ScrambleCells(df_original, y, predictors, n_permutations = 6)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{n_permutations}{# number of requested permuted datasets}
}
\description{
Suitable for ANOVA:
The function "blindDataAnova_ScrambleCellss" re-assigns at random the labels 
related to the specific combination of groups a case belongs to (e.g., 
expert: low, conflict: no, can get assigned to expert: high, conflict: yes). 
The number of permutations can be requested by user (e.g., if you want to 
have 6 datasets, input n_pertubations = 6; also defaults to 6). This function
works with unequal groups. The function returns a list with dataframes and a 
dataframe that contains all the permutations (labelled by permutation) for 
ease of use. The function also returns the number of possible permutations 
and the number of unique group combinations.
}
\keyword{ANOVA}
\keyword{Scramble}
\keyword{cells,}
