% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_ScrambleGroups.R
\name{blindDataAnova_ScrambleGroups}
\alias{blindDataAnova_ScrambleGroups}
\title{Suitable for ANOVA:
The function "blindDataAnova_ScrambleGroups" re-assigns at random the factor
levels for each predictor variable.}
\usage{
blindDataAnova_ScrambleGroups(df_original, y, predictors, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function "blindDataAnova_ScrambleGroups" re-assigns at random the factor
levels for each predictor variable.
}
\keyword{ANOVA}
\keyword{Mask}
\keyword{groups,}
