% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_ScrambleOutcome.R
\name{blindDataAnova_ScrambleOutcome}
\alias{blindDataAnova_ScrambleOutcome}
\title{Suitable for ANOVA:
The function "blindDataAnova_CreateNew" creates a new dependent variable by 
re-assigning at random the Y value of one participant to a different 
participant.}
\usage{
blindDataAnova_ScrambleOutcome(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function "blindDataAnova_CreateNew" creates a new dependent variable by 
re-assigning at random the Y value of one participant to a different 
participant.
}
\keyword{ANOVA}
\keyword{Scramble}
\keyword{outcome,}
