% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataAnova_ScramblePredictors.R
\name{blindDataAnova_ScramblePredictors}
\alias{blindDataAnova_ScramblePredictors}
\title{Suitable for ANOVA:
The function "blindDataAnova_ScramblePredictors" re-labels at random the 
grouping variables and then restores the original ordering by label name. 
Keep in mind that the factor labels (e.g., low/high) still remain, so the 
blinded dataset can still be traced back to the original variable (e.g., 
low/high is for expert variable, and conflict/no conflict is for conflict 
variable). Consider blindDataAnova_MaskGroups in combination with this 
method to prevent this.}
\usage{
blindDataAnova_ScramblePredictors(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ANOVA:
The function "blindDataAnova_ScramblePredictors" re-labels at random the 
grouping variables and then restores the original ordering by label name. 
Keep in mind that the factor labels (e.g., low/high) still remain, so the 
blinded dataset can still be traced back to the original variable (e.g., 
low/high is for expert variable, and conflict/no conflict is for conflict 
variable). Consider blindDataAnova_MaskGroups in combination with this 
method to prevent this.
}
\keyword{ANOVA}
\keyword{Scramble}
\keyword{predictors,}
