% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataRegression.R
\name{blindDataRegression}
\alias{blindDataRegression}
\title{Suitable for regression:
This function is a main function. blindDataRegression allows the user to
blind a dataset that is intended for regression}
\usage{
blindDataRegression(
  df_original,
  y,
  predictors,
  blinding_method = c("AverageWithNoisePerCase", "AddBiasToCoefficients",
    "CreateNewCoefficients", "ScrambleOutcome", "ScramblePredictors"),
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{blinding_method}{# name of blinding method to apply to the dataset
user can input any of the following blinding methods
c("AverageWithNoisePerCase",
"AddBiasToCoefficients",
"CreateNewCoefficients",
"ScrambleOutcome",
"ScramblePredictors")}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for regression:
This function is a main function. blindDataRegression allows the user to
blind a dataset that is intended for regression
}
\examples{
blindDataRegression(
df_original = data.frame(sickleave = 0:10, var1 = 100:110, var2 = 200:210),
y = "sickleave",
blinding_method = "AverageWithNoisePerCase",
predictors = c("var1", "var2")
)

}
\keyword{Regression}
\keyword{blindData,}
