% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataRegression_AddBiasToCoefficients.R
\name{blindDataRegression_AddBiasToCoefficients}
\alias{blindDataRegression_AddBiasToCoefficients}
\title{Suitable for regression:
This function adds bias to coefficients by multiplying the original
coefficient with a random value from a random uniform distribution between
(-2 and +2)}
\usage{
blindDataRegression_AddBiasToCoefficients(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for regression:
This function adds bias to coefficients by multiplying the original
coefficient with a random value from a random uniform distribution between
(-2 and +2)
}
\keyword{Add}
\keyword{Bias}
\keyword{Coefficients,}
\keyword{Regression}
\keyword{to}
