% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataRegression_AverageWithNoisePerCase.R
\name{blindDataRegression_AverageWithNoisePerCase}
\alias{blindDataRegression_AverageWithNoisePerCase}
\title{Suitable for regression:
This function average with noise per case perturbs the dependent variable
by averaging the original value with a random value from a random normal
where the mean and sd are equal to the mean and sd of the entire dataset.}
\usage{
blindDataRegression_AverageWithNoisePerCase(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for regression:
This function average with noise per case perturbs the dependent variable
by averaging the original value with a random value from a random normal
where the mean and sd are equal to the mean and sd of the entire dataset.
}
\keyword{Average}
\keyword{Regression}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{with}
