% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataRegression_CreateNewCoefficients.R
\name{blindDataRegression_CreateNewCoefficients}
\alias{blindDataRegression_CreateNewCoefficients}
\title{Suitable for regression:
This function creates new coefficients by replacing the original
coefficients with a random value from a random uniform distribution between
(-0.5 and +.5) that is multiplied by a scaling factor that is calculated by
dividing the sd of the dependent variabel from the entire data by the sd
of the predictor.}
\usage{
blindDataRegression_CreateNewCoefficients(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for regression:
This function creates new coefficients by replacing the original
coefficients with a random value from a random uniform distribution between
(-0.5 and +.5) that is multiplied by a scaling factor that is calculated by
dividing the sd of the dependent variabel from the entire data by the sd
of the predictor.
}
\keyword{Coefficients,}
\keyword{Create}
\keyword{New}
\keyword{Regression}
