% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataRegression_ScramblePredictors.R
\name{blindDataRegression_ScramblePredictors}
\alias{blindDataRegression_ScramblePredictors}
\title{Suitable for regression:
This function reassigns the predictor names at random.}
\usage{
blindDataRegression_ScramblePredictors(
  df_original,
  y,
  predictors,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for regression:
This function reassigns the predictor names at random.
}
\keyword{Regression}
\keyword{Scramble}
\keyword{predictors,}
