% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataTtest.R
\name{blindDataTtest}
\alias{blindDataTtest}
\title{Suitable for ttest:
This function is a main function. blindDataTtest allows the user to
blind a dataset that is intended for regression}
\usage{
blindDataTtest(
  df_original,
  y,
  predictor,
  blinding_method = c("AverageWithNoisePerCase", "AverageWithNoisePerGroup",
    "CenterGroups", "MaskGroups", "ScrambleGroups", "ScrambleOutcome"),
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictor}{# name of predictor, for example: "expert"}

\item{blinding_method}{# name of blinding method to apply to the dataset
user can input any of the following blinding methods
c("AverageWithNoisePerCase",
"AddBiasToCoefficients",
"CreateNewCoefficients",
"ScrambleOutcome",
"ScramblePredictors")}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
Suitable for ttest:
This function is a main function. blindDataTtest allows the user to
blind a dataset that is intended for regression
}
\examples{
blindDataTtest(
df_original = data.frame(sickleave = 1:10, var1 = as.factor(rep(c(0,1), each = 5))),
y = "sickleave",
blinding_method = "AverageWithNoisePerCase",
predictor = "var1"
)

}
\keyword{Average}
\keyword{Regression}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{with}
