% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataTtest_AverageWithNoisePerCase.R
\name{blindDataTtest_AverageWithNoisePerCase}
\alias{blindDataTtest_AverageWithNoisePerCase}
\title{Suitable for t-test:}
\usage{
blindDataTtest_AverageWithNoisePerCase(
  df_original,
  y,
  predictor,
  update_labels = TRUE
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictor}{# name of predictor, for example: "expert"}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
For each case separately this function draws 1 value from distribution of Y 
from total data and averages that random noise and the original score
}
\keyword{average}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{ttest}
\keyword{with}
