% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataTtest_CenterGroups.R
\name{blindDataTtest_CenterGroups}
\alias{blindDataTtest_CenterGroups}
\title{Suitable for t-test:}
\usage{
blindDataTtest_CenterGroups(df_original, y, predictor, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictor}{# name of predictor, for example: "expert"}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
For each group separately this function centers Y scores per group (i.e., 
substracts the mean of each group from the Y scores of each case in that group)
}
\keyword{center}
\keyword{groups,}
\keyword{ttest}
