% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataTtest_MaskGroups.R
\name{blindDataTtest_MaskGroups}
\alias{blindDataTtest_MaskGroups}
\title{Suitable for t-test:}
\usage{
blindDataTtest_MaskGroups(df_original, y, predictor, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictor}{# name of predictor, for example: "expert"}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
The function "blindDataTtest_MaskGroups" masks the label names for the 
independent variables. This blinding works also if there are more than 2 
groups, but the function will throw a message that there are more than 2 
groups. For example, the variable "expert" has two factor labels: "low/high". 
After blinding, this becomes "Condition 1/Condition 2"
}
\keyword{groups,}
\keyword{mask}
\keyword{ttest}
