% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindDataTtest_ScrambleGroups.R
\name{blindDataTtest_ScrambleGroups}
\alias{blindDataTtest_ScrambleGroups}
\title{Suitable for t-test:}
\usage{
blindDataTtest_ScrambleGroups(df_original, y, predictor, update_labels = TRUE)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictor}{# name of predictor, for example: "expert"}

\item{update_labels}{# if TRUE update labels to BLIND_[abbrevation]_[name]}
}
\description{
This function re-assigns at random the factor levels for each predictor.
}
\keyword{groups,}
\keyword{scramble}
\keyword{ttest}
