% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression.R
\name{plotRegression}
\alias{plotRegression}
\title{Suitable for regression:
This function is a main function. blindDataRegression allows the user to
blind a dataset that is intended for regression}
\usage{
plotRegression(
  df_original,
  y,
  predictors,
  plot_cooksdistance = TRUE,
  plot_crplot = TRUE,
  plot_hatvalues = TRUE,
  plot_qqplot = TRUE,
  plot_studonfitted = TRUE,
  hat_cutoff_min = 2,
  hat_cutoff_max = 3,
  studres_cutoff_min = -2,
  studres_cutoff_max = 2,
  lowess = TRUE,
  lowesscol = "blue"
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{plot_cooksdistance}{# Cook's distances plot}

\item{plot_crplot}{# component residual plot}

\item{plot_hatvalues}{# hat-values plot}

\item{plot_qqplot}{# qq norm plot#'}

\item{plot_studonfitted}{# studentized residuals on fitted values plot}

\item{hat_cutoff_min}{= 2 # horizontal line in plot at selected value x mean}

\item{hat_cutoff_max}{= 3 # horizontal line in plot at selected value x mean}

\item{studres_cutoff_min}{= -2 # value below which cases gets highlighted}

\item{studres_cutoff_max}{=  2 # value above which cases gets highlighted}

\item{lowess}{= TRUE # fit lowess line}

\item{lowesscol}{= "blue" # color of lowess line}
}
\description{
Suitable for regression:
This function is a main function. blindDataRegression allows the user to
blind a dataset that is intended for regression
}
\keyword{Average}
\keyword{Regression}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{with}
