% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression_ComponentResidual.R
\name{plotRegression_ComponentResidual}
\alias{plotRegression_ComponentResidual}
\title{Suitable for regression:
This function creates blinded component residual plots}
\usage{
plotRegression_ComponentResidual(df_original, y, predictors)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}
}
\description{
Suitable for regression:
This function creates blinded component residual plots
}
\keyword{Component}
\keyword{Regression}
\keyword{Residual}
\keyword{plot,}
