% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression_CooksDistance.R
\name{plotRegression_CooksDistance}
\alias{plotRegression_CooksDistance}
\title{Suitable for regression:
This function creates a plot for Cook's distance values}
\usage{
plotRegression_CooksDistance(df_original, y, predictors)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}
}
\description{
Suitable for regression:
This function creates a plot for Cook's distance values
}
\keyword{Cooks}
\keyword{Regression}
\keyword{distance,}
