% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression_Hatvalues.R
\name{plotRegression_Hatvalues}
\alias{plotRegression_Hatvalues}
\title{Suitable for regression:
This function creates a plot for hat-values}
\usage{
plotRegression_Hatvalues(
  df_original,
  y,
  predictors,
  hat_cutoff_min = 2,
  hat_cutoff_max = 3
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}

\item{hat_cutoff_min}{= 2 # horizontal line in plot at selected value x mean}

\item{hat_cutoff_max}{= 3 # horizontal line in plot at selected value x mean}
}
\description{
Suitable for regression:
This function creates a plot for hat-values
}
\keyword{Hatvalues,}
\keyword{Regression}
