% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression_QQ.R
\name{plotRegression_QQ}
\alias{plotRegression_QQ}
\title{Suitable for regression:
This function creates a QQ norm plot for the residuals}
\usage{
plotRegression_QQ(df_original, y, predictors)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)}
}
\description{
Suitable for regression:
This function creates a QQ norm plot for the residuals
}
\keyword{Hatvalues,}
\keyword{Regression}
