% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plotRegression_StudentizedResidualsOnFittedValues.R
\name{plotRegression_StudentizedResidualsOnFittedValues}
\alias{plotRegression_StudentizedResidualsOnFittedValues}
\title{Suitable for regression:
This function creates a plot of studentized residuals on fitted values}
\usage{
plotRegression_StudentizedResidualsOnFittedValues(
  df_original,
  y,
  predictors,
  studres_cutoff_min = -2,
  studres_cutoff_max = 2,
  lowess = TRUE,
  lowesscol = "blue"
)
}
\arguments{
\item{df_original}{# original dataframe}

\item{y}{# name of dependent variable}

\item{predictors}{# name of predictors, for example c("expert", "conflict)#'}

\item{studres_cutoff_min}{= -2 # value below which cases gets highlighted}

\item{studres_cutoff_max}{=  2 # value above which cases gets highlighted}

\item{lowess}{= TRUE # fit lowess line}

\item{lowesscol}{= "blue # color of lowess line}
}
\description{
Suitable for regression:
This function creates a plot of studentized residuals on fitted values
}
\keyword{Regression}
\keyword{plot,}
\keyword{residuals}
\keyword{studentized}
