% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData_1.R
\docType{data}
\name{SimData_1}
\alias{SimData_1}
\title{Simulated data}
\format{
A data frame with 150 observations on the following variables:
\describe{
  \item{\code{Y}}{Outcome variable. It is obtained by regression according to the previously explained formula.}
  \item{\code{A}}{This is the treatment condition variable. It is sampled from a binomial distribution.}
  \item{\code{X1}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X2}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X3}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X4}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X5}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X6}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X7}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X8}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X9}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X10}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X11}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X12}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X13}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X14}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X15}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X16}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X17}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X18}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X19}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{X20}}{Simulated covariate. It is sampled from a multivariate normal distribution.}
  \item{\code{gopt}}{Variable indicating whether the data is simulated such that there is qualitative interaction (0) or there is no qualitative interaction (1)}
 }
}
\usage{
SimData_1
}
\description{
Data generated sampling from a multivariate normal distribution using as parameters \eqn{samp} = 150,
\eqn{es} = 1.0,\eqn{J} = 20,\eqn{rho} = 0 where samp is the sample size, es is the effect size,
J is the number os covariates, and rho is the correlation between pairs of covariates.
The treatment condition (A) is sampled from a binomial distribution with \eqn{p} = 0.5 and
the treatment outcome (Y) is obtained using the following regression \eqn{  Y = 1 + (.25 x X1) + (.25 x X2) - (.25 x X5]) - (es x ((A - 1) ^ 2)) + error}
where the error is sampled from a standard normal distribution.
}
\details{
IMPORTANT: for questions about these data contact Elise Dusseldorp: elise.dusseldorp@fsw.leidenuniv.nl.
}
\keyword{datasets}
