% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataAnova.R
\name{simulateDataAnova}
\alias{simulateDataAnova}
\title{Suitable for ANOVA:
The function "simulateDataAnova" simulates experimental data for a 2x2 ANOVA
analysis. The function returns a data.frame "df_sim_anova" with three variables:  
one numeric dependent variable "score" and two independent factor variables 
"conflic" and "expert". By default, the function simulates 50 (cell.size = 50) 
cases per group; with 4 groups (i.e., 2x2 design) this yields 200 cases. This 
simulation is made to resemble a dataset from a psychological study.}
\usage{
simulateDataAnova(cell.size = 50)
}
\arguments{
\item{cell.size}{# number of cases to simulate per cell (i.e., group
combination)}
}
\value{
df_sim_anova = cell.size*4 length data.frame with variables in two 
colums:
}
\description{
Suitable for ANOVA:
The function "simulateDataAnova" simulates experimental data for a 2x2 ANOVA
analysis. The function returns a data.frame "df_sim_anova" with three variables:  
one numeric dependent variable "score" and two independent factor variables 
"conflic" and "expert". By default, the function simulates 50 (cell.size = 50) 
cases per group; with 4 groups (i.e., 2x2 design) this yields 200 cases. This 
simulation is made to resemble a dataset from a psychological study.
}
\keyword{ANOVA}
\keyword{Simulate}
\keyword{data,}
