% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataRegression.R
\name{simulateDataRegression}
\alias{simulateDataRegression}
\title{Suitable for regression:
The function "simulateDataRegression" simulates correlational data for
regression analysis. The function returns a data.frame "df_sim_reg" with
five numeric variables:  one dependent variable "sickleave" (interval) and
four independent variables "gender", "general_health", "stress_at_work", and
"var_of_work_ac" (abbr. variety of work activities). By default, the function
simulates 85 (n = 85) cases. This simulation is made to resemble a dataset
from a psychological study.}
\usage{
simulateDataRegression(n = 85)
}
\arguments{
\item{n}{# number of cases to simulate}
}
\description{
Suitable for regression:
The function "simulateDataRegression" simulates correlational data for
regression analysis. The function returns a data.frame "df_sim_reg" with
five numeric variables:  one dependent variable "sickleave" (interval) and
four independent variables "gender", "general_health", "stress_at_work", and
"var_of_work_ac" (abbr. variety of work activities). By default, the function
simulates 85 (n = 85) cases. This simulation is made to resemble a dataset
from a psychological study.
}
\keyword{Average}
\keyword{Regression}
\keyword{case,}
\keyword{noise}
\keyword{per}
\keyword{with}
