% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataTtest.R
\name{simulateDataTtest}
\alias{simulateDataTtest}
\title{Suitable for t-test:
The function "simulateDataRegression" simulates correlational simulates 
experimental data for a 2 group t-test analysis. The function returns a 
data.frame "df_sim_ttest" with two variables: one numeric dependent 
variable "score" and one independent factor variables "expert". By default,
the function simulates 50 (cell.size = 50) cases per group; with 2 groups 
this yields 100 cases. This simulation is made to resemble a dataset from
a psychological study. 
from a psychological study.}
\usage{
simulateDataTtest(cell.size = 50)
}
\arguments{
\item{cell.size}{# number of cases to simulate per cell (i.e., group
combination)}
}
\value{
df_sim_ttest = cell.size*4 length data.frame with variables in two 
colums:
}
\description{
Suitable for t-test:
The function "simulateDataRegression" simulates correlational simulates 
experimental data for a 2 group t-test analysis. The function returns a 
data.frame "df_sim_ttest" with two variables: one numeric dependent 
variable "score" and one independent factor variables "expert". By default,
the function simulates 50 (cell.size = 50) cases per group; with 2 groups 
this yields 100 cases. This simulation is made to resemble a dataset from
a psychological study. 
from a psychological study.
}
\keyword{Simulate}
\keyword{data,}
\keyword{ttest}
